requirements_minimal.txt
# Core ML/DL packages
# torch==2.4.0
# torchvision==0.19.0
# torchaudio==2.4.0
pytorch-lightning==2.5.6
lightning-utilities

# Data processing
numpy==1.24.4
pandas
scipy
scikit-learn
scikit-image
opencv-python==4.11.0.86
albumentations==2.0.5
monai
SimpleITK==2.4.0

# Hydra and configuration
hydra-core==1.3.1
hydra-colorlog==1.2.0
omegaconf==2.3.0

# Logging and monitoring
wandb
tensorboardX==2.6.2.2

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Model utilities
timm==1.0.15
torchmetrics
segmentation_models_pytorch==0.4.0
efficientnet_pytorch==0.7.1

# Other utilities
tqdm
PyYAML
requests
pillow
einops==0.8.0
pyrootutils==1.0.4